+++ Frontend/src/pages/crowdactions/Detail/CrowdactionDetails.tsx
@@ -26,7 +26,7 @@ import Formatter from '../../../formatter';
 import LazyImage from '../../../components/LazyImage/LazyImage';
 import { TextField } from 'formik-material-ui';
 import CrowdactionComments from '../../../components/CrowdactionComments/CrowdactionComments';
-import SocialMediaSharing from './SocialMediaSharing';
+import SocialMediaSharing from '../SocialMediaSharing/SocialMediaSharing';
 
 type TParams = {
   slug: string;
